@@ -2,8 +2,8 @@ import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 
 class DashboardEntity extends Equatable {
-  final String title;
-  final String image;
+  final String? title;
+  final String? image;
 
   DashboardEntity({
     @required this.title,
@@ -11,5 +11,5 @@ class DashboardEntity extends Equatable {
   });
   @override
   // TODO: implement props
-  List<Object> get props => [title, image];
+  List<Object> get props => [title!, image!];
 }
