@@ -1,42 +1,43 @@
-// ignore_for_file: file_names
+// ignore_for_file: file_names, prefer_const_constructors
 
 import 'package:flutter/material.dart';
 
 import '../constants/colors.dart';
 import '../constants/widgetFunctions.dart';
 
-AppBar appBar(String text, bool isDashboard, bool isIndex,
-    {BuildContext context}) {
+AppBar appBar(String? text, bool? isDashboard, bool? isIndex,
+    {BuildContext? context}) {
   return AppBar(
     backgroundColor: AZURE,
     elevation: 0.5,
-    leading: isDashboard
+    leading: isDashboard!
         ? Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 7.5),
+            padding: EdgeInsets.symmetric(horizontal: 7.5),
             child: Image.asset(
               'assets/images/logo.png',
               height: 30,
               filterQuality: FilterQuality.high,
             ),
           )
-        : {
-            (isIndex)
+        : 
+            (isIndex!)
                 ? null
                 : IconButton(
-                    onPressed: () => Navigator.pop(context),
+                    onPressed: () => Navigator.pop(context!),
                     icon: const Icon(
                       Icons.navigate_before,
                       color: Colors.black,
                       size: 30,
                     ),
                   ),
-          },
+          
     centerTitle: true,
-    title: addSubText(text, fontStyle: FontStyle.normal),
+    title: addSubText(text!, fontStyle: FontStyle.normal),
   );
 }
 
-AppBar appBarNoLogo(String text, bool isIndex, Color color,{BuildContext context}) {
+AppBar appBarNoLogo(String text, bool isIndex, Color color,
+    {BuildContext? context}) {
   return AppBar(
     backgroundColor: AZURE,
     elevation: 0.5,
@@ -44,7 +45,7 @@ AppBar appBarNoLogo(String text, bool isIndex, Color color,{BuildContext context
     leading: isIndex
         ? null
         : IconButton(
-            onPressed: () => Navigator.pop(context),
+            onPressed: () => Navigator.pop(context!),
             icon: Icon(
               Icons.navigate_before,
               color: color,
